test: split refresh coalescing coverage
